Move auth functionality into separate package
