+++ docs/configuration.md
@@ -154,7 +154,7 @@ The following subsections cover all CLI flags which promscale supports. You can 
 | web.auth.password          | string  |      ""       | Authentication password used for web endpoint authentication. This flag should be set together with auth-username. It is mutually exclusive with auth-password-file and bearer-token methods.                               |
 | web.auth.password-file     | string  |      ""       | Path for auth password file containing the actual password used for web endpoint authentication. This flag should be set together with auth-username. It is mutually exclusive with auth-password and bearer-token methods. |
 | web.auth.username          | string  |      ""       | Authentication username used for web endpoint authentication. Disabled by default.                                                                                                                                          |
-| web.auth.ignore-path       | string  |      ""       | Http paths which has to be skipped from authentication. This flag shall be repeated and each one would be appended to the ignore list.                                                                                      |
+| web.auth.ignore-path       | string  |      ""       | HTTP paths which has to be skipped from authentication. This flag shall be repeated and each one would be appended to the ignore list.                                                                                      |
 | web.cors-origin            | string  |     `.*`      | Regex for CORS origin. It is fully anchored. Example: 'https?://(domain1                                                                                                                                                    |
 | web.enable-admin-api       | boolean |     false     | Allow operations via API that are for advanced users. Currently, these operations are limited to deletion of series.                                                                                                        |
 | web.listen-address         | string  |    `:9201`    | Address to listen on for web endpoints.                                                                                                                                                                                     |
