+++ pkg/api/common_test.go
@@ -6,11 +6,8 @@ package api
 
 import (
 	"context"
-	"errors"
 	"net/http"
 	"net/http/httptest"
-	"os"
-	"path"
 	"reflect"
 	"strings"
 	"testing"
@@ -120,160 +117,6 @@ func doCORSWrapperRequest(t *testing.T, queryHandler http.Handler, url, origin s
 	w := httptest.NewRecorder()
 	queryHandler.ServeHTTP(w, req)
 	return w
-}
-
-func TestValidateConfig(t *testing.T) {
-	fileContents, err := os.ReadFile("common_test.go")
-	if err != nil {
-		t.Fatal("error reading file contents common_test.go")
-	}
-	fileContentsString := strings.TrimSpace(string(fileContents))
-	testCases := []struct {
-		name      string
-		cfg       *Auth
-		returnErr error
-		passSet   string
-		tokenSet  string
-	}{
-		{
-			name: "empty config",
-			cfg:  &Auth{},
-		},
-		{
-			name: "basic auth and bearer token set",
-			cfg: &Auth{
-				BasicAuthUsername: "foo",
-				BearerToken:       "foo",
-			},
-			returnErr: usernameAndTokenFlagsSetError,
-		},
-		{
-			name: "basic auth missing password",
-			cfg: &Auth{
-				BasicAuthUsername: "foo",
-			},
-			returnErr: noPasswordFlagsSetError,
-		},
-		{
-			name: "basic auth password and password file set",
-			cfg: &Auth{
-				BasicAuthUsername:     "foo",
-				BasicAuthPassword:     "foo",
-				BasicAuthPasswordFile: "foo",
-			},
-			returnErr: multiplePasswordFlagsSetError,
-		},
-		{
-			name: "basic auth invalid password file",
-			cfg: &Auth{
-				BasicAuthUsername:     "foo",
-				BasicAuthPasswordFile: "invalid filename",
-			},
-			returnErr: os.ErrNotExist,
-		},
-		{
-			name: "basic auth password set",
-			cfg: &Auth{
-				BasicAuthUsername: "foo",
-				BasicAuthPassword: "pass",
-			},
-			passSet: "pass",
-		},
-		{
-			name: "basic auth no username set",
-			cfg: &Auth{
-				BasicAuthPassword: "pass",
-			},
-			returnErr: noUsernameFlagSetError,
-		},
-		{
-			name: "basic auth password file set",
-			cfg: &Auth{
-				BasicAuthUsername:     "foo",
-				BasicAuthPasswordFile: "common_test.go",
-			},
-			passSet: fileContentsString,
-		},
-		{
-			name: "bearer token and token file set",
-			cfg: &Auth{
-				BearerToken:     "foo",
-				BearerTokenFile: "foo",
-			},
-			returnErr: multipleTokenFlagsSetError,
-		},
-		{
-			name: "bearer token set",
-			cfg: &Auth{
-				BearerToken: "foo",
-			},
-			tokenSet: "foo",
-		},
-		{
-			name: "bearer token file set",
-			cfg: &Auth{
-				BearerTokenFile: "common_test.go",
-			},
-			tokenSet: fileContentsString,
-		},
-		{
-			name: "bearer token file invalid file set",
-			cfg: &Auth{
-				BearerTokenFile: "invalid file",
-			},
-			returnErr: os.ErrNotExist,
-		},
-		{
-			name: "all config options set",
-			cfg: &Auth{
-				BasicAuthUsername:     "set",
-				BasicAuthPassword:     "set",
-				BasicAuthPasswordFile: "set",
-				BearerToken:           "set",
-				BearerTokenFile:       "set",
-			},
-			returnErr: usernameAndTokenFlagsSetError,
-		},
-		{
-			name: "invalid ignore path",
-			cfg: &Auth{
-				IgnorePaths: []string{
-					"[",
-				},
-			},
-			returnErr: path.ErrBadPattern,
-		},
-		{
-			name: "valid ignore path",
-			cfg: &Auth{
-				IgnorePaths: []string{
-					"/hello",
-				},
-			},
-		},
-	}
-
-	for _, c := range testCases {
-		t.Run(c.name, func(t *testing.T) {
-			err := Validate(&Config{
-				Auth: c.cfg,
-			})
-			if c.returnErr != nil {
-				if !errors.Is(err, c.returnErr) {
-					t.Errorf("unexpected error received: %s", err)
-				}
-			} else if err != nil {
-				t.Errorf("unexpected error received: %s", err)
-			}
-
-			if c.passSet != "" && c.cfg.BasicAuthPassword != c.passSet {
-				t.Errorf("unexpected password set: got %s wanted %s", c.cfg.BasicAuthPassword, c.passSet)
-			}
-			if c.tokenSet != "" && c.cfg.BearerToken != c.tokenSet {
-				t.Errorf("unexpected bearer token set: got %s wanted %s", c.cfg.BearerToken, c.tokenSet)
-			}
-		})
-	}
 }
 
 func TestMarshalExemplar(t *testing.T) {
