+++ pkg/tests/end_to_end_tests/db_connections_test.go
@@ -21,7 +21,7 @@ func TestDBConnectionHandling(t *testing.T) {
 	withDB(t, *testDatabase, func(dbOwner *pgxpool.Pool, tb testing.TB) {
 		db := testhelpers.PgxPoolWithRole(t, *testDatabase, "prom_writer")
 		defer db.Close()
-		_, pgClient, err := buildRouterWithAPIConfig(db, defaultAPIConfig())
+		_, pgClient, err := buildRouterWithAPIConfig(db, defaultAPIConfig(), nil)
 		if err != nil {
 			t.Fatalf("unexpected error while creating pgClient: %s", err)
 		}
