+++ pkg/tests/end_to_end_tests/promql_write_endpoint_test.go
@@ -185,7 +185,7 @@ func verifyTimeseries(t testing.TB, db *pgxpool.Pool, tsSlice []prompb.TimeSerie
 }
 
 func sendConcurrentWrites(t testing.TB, db *pgxpool.Pool, queues int, metricGroups int, totalRequests int, duplicates bool) {
-	router, pgClient, err := buildRouterWithAPIConfig(db, defaultAPIConfig())
+	router, pgClient, err := buildRouterWithAPIConfig(db, defaultAPIConfig(), nil)
 	if err != nil {
 		t.Fatalf("Unable to send concurrent writes, error building router: %s", err)
 	}
